[ds-fusion] Move the resource requests class into a dedicated header
This patch moves the resource requests class outside the gpu_executable.
This is required because it will be used for thunk initializations in
testing thunks. Also renamed the virtual class to
`ResourceRequestsInterface` to avoid confusion with the concrete class.
