+++ xla/service/gpu/runtime/command_buffer_thunk.h
@@ -45,7 +45,7 @@ class CommandBufferThunk : public Thunk {
   const std::unique_ptr<SequentialThunk>& thunks() const { return thunks_; }
 
   absl::Status Prepare(const PrepareParams& params,
-                       ResourceRequests& resource_requests) override;
+                       ResourceRequestsInterface& resource_requests) override;
   absl::Status Initialize(const InitializeParams& params) override;
   absl::Status ExecuteOnStream(const ExecuteParams& params) override;
