+++ xla/service/gpu/runtime/custom_call_thunk.h
@@ -82,7 +82,7 @@ class CustomCallThunk : public Thunk {
       const HloComputation* called_computation);
 
   absl::Status Prepare(const PrepareParams& params,
-                       ResourceRequests& resource_requests) override;
+                       ResourceRequestsInterface& resource_requests) override;
   absl::Status Initialize(const InitializeParams& params) override;
   absl::Status ExecuteOnStream(const ExecuteParams& params) override;
