+++ xla/service/gpu/runtime/dynamic_slice_thunk.h
@@ -65,7 +65,7 @@ class DynamicSliceThunk : public Thunk {
   const Thunk* embedded_thunk() const { return embedded_thunk_.get(); }
 
   absl::Status Prepare(const PrepareParams& params,
-                       ResourceRequests& resource_requests) override;
+                       ResourceRequestsInterface& resource_requests) override;
   absl::Status Initialize(const InitializeParams& params) override;
   absl::Status ExecuteOnStream(const ExecuteParams& params) override;
