+++ xla/service/gpu/runtime/sequential_thunk.h
@@ -39,7 +39,7 @@ class SequentialThunk : public Thunk {
   std::string ToString(int indent) const override;
 
   absl::Status Prepare(const PrepareParams& params,
-                       ResourceRequests& resource_requests) override;
+                       ResourceRequestsInterface& resource_requests) override;
   absl::Status Initialize(const InitializeParams& params) override;
   absl::Status ExecuteOnStream(const ExecuteParams& params) override;
